refactor(app-shell): extract useShellConnections
Move the LLM-connection cluster — the connection list, default-connection
slug, connectionsEqual dedup, refreshConnections, and handleConnectionEvent
— into a dedicated hook. setConnections/setDefaultConnection are returned so
the onboarding-snapshot seed (which stays in AppShell to also seed sessions)
can prime them. Pure move, zero behavior change.

Partial vs the map's blade 5: themePref/themePalette, userLabel and
defaultPermissionMode stay in app-shell.tsx — the default-permission-mode
contract pins the defaultPermissionMode useState + refreshShellSettings +
closeSettings mirror to app-shell.tsx specifically, and the theme setters
have multiple app-shell writers (visual-smoke, settings-overlay onChange),
so moving them would require setter injection with no net simplification.

Blade 5 of Round B app-shell decomposition.
